Move diagnostics to lib
