+++ CMakeLists.txt
@@ -195,7 +195,6 @@ add_subdirectory(bftengine)
 add_subdirectory(tools)
 add_subdirectory(storage)
 add_subdirectory(scripts)
-add_subdirectory(diagnostics)
 add_subdirectory(client)
 if(USE_GRPC)
 	add_subdirectory(thin-replica-server)
