+++ bftengine/src/bcstatetransfer/BCStateTran.hpp
@@ -36,8 +36,8 @@
 #include "util/Handoff.hpp"
 #include "SysConsts.hpp"
 #include "util/throughput.hpp"
-#include "diagnostics.h"
-#include "performance_handler.h"
+#include "diagnostics.hpp"
+#include "performance_handler.hpp"
 #include "util/Timers.hpp"
 #include "TimeUtils.hpp"
 #include "util/SimpleMemoryPool.hpp"
