+++ bftengine/src/bftengine/ReplicaImp.hpp
@@ -31,8 +31,8 @@
 #include "util/OpenTracing.hpp"
 #include "RequestHandler.h"
 #include "InternalBFTClient.hpp"
-#include "diagnostics.h"
-#include "performance_handler.h"
+#include "diagnostics.hpp"
+#include "performance_handler.hpp"
 #include "RequestsBatchingLogic.hpp"
 #include "ReplicaStatusHandlers.hpp"
 #include "PerformanceManager.hpp"
