+++ bftengine/src/bftengine/ReplicaStatusHandlers.cpp
@@ -10,7 +10,7 @@
 // file.
 
 #include "ReplicaStatusHandlers.hpp"
-#include "diagnostics.h"
+#include "diagnostics.hpp"
 #include "util/json_output.hpp"
 
 #include <string>
