+++ kvbc/src/categorization/kv_blockchain.cpp
@@ -14,8 +14,8 @@
 #include "categorization/kv_blockchain.h"
 #include "bcstatetransfer/SimpleBCStateTransfer.hpp"
 #include "bftengine/ControlStateManager.hpp"
-#include "diagnostics.h"
-#include "performance_handler.h"
+#include "diagnostics.hpp"
+#include "performance_handler.hpp"
 #include "kvbc_key_types.hpp"
 #include "categorization/db_categories.h"
 #include "util/endianness.hpp"
