+++ libs/CMakeLists.txt
@@ -6,5 +6,5 @@ add_subdirectory(crypto)
 add_subdirectory(secrets)
 add_subdirectory(communication)
 add_subdirectory(reconfiguration)
-
+add_subdirectory(diagnostics)
