+++ libs/diagnostics/README.md
