+++ libs/diagnostics/performance_handler.hpp
