+++ libs/diagnostics/src/performance_handler.cpp
@@ -13,7 +13,7 @@
 
 #include <iomanip>
 
-#include "performance_handler.h"
+#include "performance_handler.hpp"
 
 static logging::Logger DIAG_LOGGER = logging::getLogger("concord.diag.perf");
