+++ libs/diagnostics/status_handlers.hpp
