+++ libs/diagnostics/test/CMakeLists.txt
